FE: Create useThemeState and useThemeContext hooks
